Added bullets and modified enemy location storage
Shifted the grid calculators to - TILE_SIZE / 2 cuz of how its 1 smller
on each side of the grid. Added bullets from towers.

+++ src/entity/Entity.java
@@ -6,7 +6,7 @@ public abstract class Entity {
 	
 	double x, y;
 	
-	public Entity(int x, int y) {
+	public Entity(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
